Sales order, sales invoice. Sales order half

+++ SriLankanParadise.ERP.UserManagement/Business Service/LocationInventoryService.cs
@@ -93,7 +93,7 @@ namespace SriLankanParadise.ERP.UserManagement.Business_Service
             return await _locationInventoryRepository.GetLowStockItemsByLocationOnly(locationId);
         }
 
-        public async Task ReduceInventoryByFIFO(int locationId, int itemMasterId, int quantity)
+        public async Task ReduceInventoryByFIFO(int locationId, int itemMasterId, decimal quantity)
         {
             await _locationInventoryRepository.ReduceInventoryByFIFO(locationId, itemMasterId, quantity);
         }
