+++ SriLankanParadise.ERP.UserManagement/ERP Web/DTOs/LocationInventoryMovementDto.cs
@@ -18,6 +18,6 @@
 
         public DateTime? Date { get; set; }
 
-        public int? Qty { get; set; }
+        public decimal? Qty { get; set; }
     }
 }
